requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
dotenv==0.9.9
idna==3.11
mysql-connector-python==9.7.0
pycountry==26.2.16
python-dotenv==1.2.2
requests==2.33.1
urllib3==2.6.3
